+++ packages/shared/common/src/index.ts
@@ -2,8 +2,12 @@ import AttributeReference from './AttributeReference';
 import Context from './Context';
 import ContextFilter from './ContextFilter';
 import {
+  Backoff,
+  CompositeDataSource,
   DataSourceErrorKind,
+  DefaultBackoff,
   LDFileDataSourceError,
+  LDFlagDeliveryFallbackError,
   LDPollingError,
   LDStreamingError,
   StreamingErrorHandler,
@@ -22,9 +26,13 @@ export {
   AttributeReference,
   Context,
   ContextFilter,
+  CompositeDataSource,
   DataSourceErrorKind,
+  Backoff,
+  DefaultBackoff,
   LDPollingError,
   LDStreamingError,
   StreamingErrorHandler,
   LDFileDataSourceError,
+  LDFlagDeliveryFallbackError,
 };