+++ packages/tooling/jest/CHANGELOG.md
@@ -1,5 +1,59 @@
 # Changelog
 
+## [0.1.7](https://github.com/launchdarkly/js-core/compare/jest-v0.1.6...jest-v0.1.7) (2025-05-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.10.1 to ~10.10.2
+
+## [0.1.6](https://github.com/launchdarkly/js-core/compare/jest-v0.1.5...jest-v0.1.6) (2025-05-21)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.10.0 to ~10.10.1
+
+## [0.1.5](https://github.com/launchdarkly/js-core/compare/jest-v0.1.4...jest-v0.1.5) (2025-04-29)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.9.9 to ~10.10.0
+
+## [0.1.4](https://github.com/launchdarkly/js-core/compare/jest-v0.1.3...jest-v0.1.4) (2025-04-16)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.9.8 to ~10.9.9
+
+## [0.1.3](https://github.com/launchdarkly/js-core/compare/jest-v0.1.2...jest-v0.1.3) (2025-04-08)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.9.7 to ~10.9.8
+
+## [0.1.2](https://github.com/launchdarkly/js-core/compare/jest-v0.1.1...jest-v0.1.2) (2025-03-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/react-native-client-sdk bumped from ~10.9.6 to ~10.9.7
+
 ## [0.1.1](https://github.com/launchdarkly/js-core/compare/jest-v0.1.0...jest-v0.1.1) (2025-02-06)